Allow configuration of db path
The database path is no longer hard-coded in the code. Instead, the user
is required to specify the database path in the call with `--dbpath`.
Alternatively, the dbpath can be saved with `--setdbpath` so that then
it will not be required during subsequent invocations. The stored path
can also be removed with `--remove_dbpath` if this retention is no
longer desired.